requirements.txt
matplotlib==3.8.2
missingno==0.5.2
numba==0.58.1
numpy==1.26.2
openpyxl==3.1.2
pandas==2.1.3
Pillow==10.1.0
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.0
shap==0.43.0
tqdm==4.66.1
tzdata==2023.3
umap-learn==0.5.5
GEOparse-2.0.3
click-8.1.7
pyyaml-6.0.1